Added retroalimentation to probabilistic model

+++ core/probabilistic/ProbabilisticQueryProcessor.py
@@ -1,4 +1,5 @@
 from core.basics.BasicQueryProcessor import BasicQueryProcessor
+from numpy import float32
 
 
 class ProbabilisticQueryProcessor(BasicQueryProcessor):
@@ -7,4 +8,4 @@ class ProbabilisticQueryProcessor(BasicQueryProcessor):
 
     def ProcessQuery(self, query):
         query.tokens = self.vectorizer.CountTransform(str(query))
-        return query.tokens
+        return query.tokens.astype(float32)
